requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
django-heroku==0.3.1
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
gunicorn==20.1.0
mysqlclient==2.1.0
Pillow==9.0.1
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
whitenoise==6.0.0
